Convolutional networks done

+++ assignment2/cs231n/im2col.py
@@ -7,8 +7,8 @@ def get_im2col_indices(x_shape, field_height, field_width, padding=1, stride=1):
     N, C, H, W = x_shape
     assert (H + 2 * padding - field_height) % stride == 0
     assert (W + 2 * padding - field_height) % stride == 0
-    out_height = (H + 2 * padding - field_height) / stride + 1
-    out_width = (W + 2 * padding - field_width) / stride + 1
+    out_height = int((H + 2 * padding - field_height) / stride) + 1
+    out_width = int((W + 2 * padding - field_width) / stride) + 1
 
     i0 = np.repeat(np.arange(field_height), field_width)
     i0 = np.tile(i0, C)
@@ -28,7 +28,7 @@ def im2col_indices(x, field_height, field_width, padding=1, stride=1):
     # Zero-pad the input
     p = padding
     x_padded = np.pad(x, ((0, 0), (0, 0), (p, p), (p, p)), mode='constant')
-
+    #print("in first function",stride)
     k, i, j = get_im2col_indices(x.shape, field_height, field_width, padding,
                                  stride)
